added some docs and comments

+++ LanceOfDestiny-FourthDimension/src/main/java/org/LanceOfDestiny/ui/BuildModeController.java
@@ -32,7 +32,6 @@ public class BuildModeController {
 
 
     //Creates obstacles(from given input) and places them in a list where they will be stored and shuffled for randomization
-    //todo: obstacleları ekleyebileceğimiz bir listeye ihtiyacımız var gameobjesinde olabilir
     public void addBarriers(int simpNum, int firmNum, int expNum, int giftNum) {
 
         int i;
@@ -62,6 +61,9 @@ public class BuildModeController {
         LanceOfDestiny.getInstance().setGameMap(new GameMap());
     }
 
+    // todo: burada gapler daha sonra tekrardan kontrol edilmeli çünkü
+    //  control panel'ın ekrandaki konumu bu değerleri etkileyecek mi emin değilim
+
     public void intializeBarrierCoordinates() {
         barrierList = LanceOfDestiny.getInstance().getGameMap().getBarriers();
         if (barrierList.isEmpty()) return;
@@ -71,7 +73,7 @@ public class BuildModeController {
         int barrierAndGapPairs = 1;
         float verticalGap = 0.0F;
         for(int i =1; i < barrierList.size(); i++) {
-            double widthCheck = barrierAndGapWidth * barrierAndGapPairs + 40;
+            float widthCheck = barrierAndGapWidth * barrierAndGapPairs + 40F;
             Barrier nextBarrier = barrierList.get(i);
             if (widthCheck + barrierAndGapWidth > 1280) {
                 barrierAndGapPairs = 1;
